Implemented toggle view for every tab in the info window view.
