+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-view.js
@@ -2,7 +2,6 @@ var cdb = require('cartodb.js');
 var createTextLabelsTabPane = require('../../components/tab-pane/create-text-labels-tab-pane');
 var TabPaneTemplate = require('./layer-tab-pane.tpl');
 var Header = require('./layer-header-view.js');
-var DataView = require('./layer-content-views/data/data-view');
 var AnalysesView = require('./layer-content-views/analyses-view');
 var InfowindowView = require('./layer-content-views/infowindow-view');
