+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/infowindow-view.js
@@ -1,40 +1,106 @@
 var cdb = require('cartodb.js');
-var EditionTogglePanelView = require('../../../components/edition-toggle/edition-toggle-panel-view');
-var CodeMirrorView = require('../../../components/code-mirror/code-mirror-view');
 var InfowindowView = require('./infowindow/infowindow-view');
+var TooltipView = require('./infowindow/infowindow-view');
+var createTextLabelsTabPane = require('../../../components/tab-pane/create-text-labels-tab-pane');
+var TabPaneTemplate = require('../../tab-pane-submenu.tpl');
 
 module.exports = cdb.core.View.extend({
+
+  initialize: function (opts) {
+    if (!opts.layerDefinitionModel) throw new Error('Layer definition is required');
+    if (!opts.querySchemaModel) throw new Error('querySchemaModel is required');
+    this._layerDefinitionModel = opts.layerDefinitionModel;
+    this._querySchemaModel = opts.querySchemaModel;
+    this._layerInfowindowModel = this._layerDefinitionModel.infowindowModel;
+    this._layerTooltipModel = this._layerDefinitionModel.tooltipModel;
+  },
+
   render: function () {
+    this.clearSubViews();
+    this.$el.empty();
+
+    this._initViews();
+
+    return this;
+  },
+
+  _initViews: function () {
     var self = this;
 
-    // CodeMirror placeholder
-    var CodeMirrorModel = new cdb.core.Model({
-      content: 'Hola mundo'
-    });
-
     var tabPaneTabs = [{
-      label: 'VALUES',
-      selected: true,
+      label: _t('editor.layers.infowindow-menu-tab-pane-labels.click'),
       createContentView: function () {
-        return new InfowindowView(self.options);
+        return new InfowindowView({
+          querySchemaModel: self._querySchemaModel,
+          layerInfowindowModel: self._layerInfowindowModel,
+          layerDefinitionModel: self._layerDefinitionModel,
+          toggleLabels: {
+            off: 'VALUES',
+            on: 'HTML'
+          },
+          templateStyles: [
+            {
+              value: 'none',
+              label: _t('editor.layers.infowindow.style.none')
+            }, {
+              value: 'infowindow_light',
+              label: _t('editor.layers.infowindow.style.light')
+            }, {
+              value: 'infowindow_dark',
+              label: _t('editor.layers.infowindow.style.dark')
+            }, {
+              value: 'infowindow_light_header_blue',
+              label: _t('editor.layers.infowindow.style.color')
+            }, {
+              value: 'infowindow_header_with_image',
+              label: _t('editor.layers.infowindow.style.image')
+            }
+          ]
+        });
       }
     }, {
-      label: 'HTML',
-      selected: false,
+      label: _t('editor.layers.infowindow-menu-tab-pane-labels.hover'),
       createContentView: function () {
-        return new CodeMirrorView({
-          model: CodeMirrorModel
+        return new TooltipView({
+          querySchemaModel: self._querySchemaModel,
+          layerInfowindowModel: self._layerTooltipModel,
+          layerDefinitionModel: self._layerDefinitionModel,
+          toggleLabels: {
+            off: 'VISUAL',
+            on: 'MANUAL'
+          },
+          templateStyles: [
+            {
+              value: 'none',
+              label: _t('editor.layers.infowindow.style.none')
+            }, {
+              value: 'tooltip_dark',
+              label: _t('editor.layers.infowindow.style.dark')
+            }, {
+              value: 'tooltip_light',
+              label: _t('editor.layers.infowindow.style.light')
+            }
+          ]
         });
       }
     }];
 
-    var editionTogglePanelView = new EditionTogglePanelView({
-      panes: tabPaneTabs
-    });
+    var tabPaneOptions = {
+      tabPaneOptions: {
+        template: TabPaneTemplate,
+        tabPaneItemOptions: {
+          tagName: 'li',
+          className: 'CDB-NavSubmenu-item'
+        }
+      },
+      tabPaneItemLabelOptions: {
+        tagName: 'button',
+        className: 'CDB-NavSubmenu-link u-upperCase'
+      }
+    };
 
-    this.$el.append(editionTogglePanelView.render().el);
-    this.addView(editionTogglePanelView);
-
-    return this;
+    this._layerTabPaneView = createTextLabelsTabPane(tabPaneTabs, tabPaneOptions);
+    this.$el.append(this._layerTabPaneView.render().$el);
+    this.addView(this._layerTabPaneView);
   }
 });
